requirements.txt
pypdf
python-dotenv==1.0.1
langchain
chromadb # Vector storage
pytest
boto3
typing-inspect==0.8.0
typing_extensions==4.5.0
pydantic==1.10.8
xlrd
langchain-openai==0.1.8 # For embeddings